Resolve signedness warning.

+++ Machines/Enterprise/Enterprise.cpp
@@ -629,7 +629,7 @@ private:
 		const auto ram_floor = (0x100 << 14) - ram_.size();
 			// Each segment is 2^14 bytes long and there are 256 of them. So the Enterprise has a 22-bit address space.
 			// RAM is at the end of that range; `ram_floor` is the 22-bit address at which RAM starts.
-		return &ram_[(page << 14) - ram_floor];
+		return &ram_[size_t((page << 14) - ram_floor)];
 	}
 
 	const uint8_t *read_pointers_[4] = {nullptr, nullptr, nullptr, nullptr};
